requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-compress==1.14
gunicorn==21.2.0

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1
pdfminer.six==20221105

# OCR for scanned PDFs (REQUIRED for real estate documents)
pdf2image==1.16.3
pytesseract==0.3.10
Pillow>=10.3.0  # Use latest compatible version for Python 3.13

# Google Vision API - Cloud OCR (fast, accurate, cost-effective) - PRIMARY OCR
google-cloud-vision==3.7.2

# PaddleOCR - DISABLED to save memory (using Google Vision instead)
# paddleocr==2.7.3
# paddlepaddle==2.6.2  # CPU version
# numpy<2.0  # Required by PaddleOCR

# NumPy - Still needed by other packages
numpy>=1.24.0,<2.0  # Compatible with scikit-learn

# Authentication & Database
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Limiter==3.5.0  # Rate limiting for security
authlib==1.3.0
requests==2.31.0
Werkzeug==3.0.1

# Payment Processing
stripe==7.9.0

# AI/ML - CRITICAL
anthropic>=0.39.0  # Updated for latest API compatibility
scikit-learn==1.3.2  # For Risk DNA similarity matching

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
psutil==5.9.6  # Memory monitoring for debugging

# Email
resend==2.0.0  # Transactional emails
